removed attester code. builds
