+++ node/components/network/src/gossip/fetch.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeMap;
 
 use anyhow::Context as _;
 use zksync_concurrency::{ctx, oneshot, scope, sync};
-use zksync_consensus_roles::{attester, validator};
+use zksync_consensus_roles::validator;
 use zksync_consensus_storage::BlockStoreState;
 
 /// A block fetching request.
