+++ node/components/network/src/proto/gossip.proto
@@ -4,7 +4,6 @@ package zksync.network.gossip;
 
 import "zksync/roles/node.proto";
 import "zksync/roles/validator.proto";
-import "zksync/roles/attester.proto";
 
 // First message exchanged in the encrypted session.
 message Handshake {
@@ -19,19 +18,6 @@ message Handshake {
 message PushValidatorAddrs {
   // Signed roles.validator.Msg.net_address.
   repeated roles.validator.Signed net_addresses = 1;
-}
-
-message PushBatchVotes {
-  // Requesting the peer to respond with votes for the batch. 
-  optional roles.attester.Batch want_votes_for = 2; // optional
-  // Signed roles.validator.Msg.votes
-  repeated roles.attester.Signed votes = 1;
-}
-
-message PushBatchVotesResp {
-  // Signed roles.validator.Msg.votes
-  // Empty if want_votes_for in request was not set.
-  repeated roles.attester.Signed votes = 1;
 }
 
 message Last {
