+++ node/components/network/src/rpc/mod.rs
@@ -28,7 +28,6 @@ pub(crate) mod consensus;
 pub(crate) mod get_block;
 mod metrics;
 pub(crate) mod ping;
-pub(crate) mod push_batch_votes;
 pub(crate) mod push_block_store_state;
 pub(crate) mod push_validator_addrs;
 #[cfg(test)]
