--- node/libs/roles/src/attester/messages/mod.rs
@@ -1,8 +1,0 @@
-//! Attester messages.
-mod batch;
-mod msg;
-
-pub use batch::*;
-pub use msg::*;
-
-pub use crate::validator::GenesisHash;
