--- node/libs/roles/src/attester/mod.rs
@@ -1,11 +1,0 @@
-//! Attester role implementation.
-
-#[cfg(test)]
-mod tests;
-
-mod conv;
-mod keys;
-mod messages;
-mod testonly;
-
-pub use self::{keys::*, messages::*};
