+++ node/libs/roles/src/lib.rs
@@ -6,10 +6,7 @@
 //! - `Validator`: a node that participates in the consensus protocol, so it votes for blocks and produces blocks.
 //!   It also participates in the validator network, which is a mesh network just for validators. Not
 //!   every node has this role.
-//! - `Attester`: a node that signs the L1 batches and broadcasts the signatures known as votes to the gossip network.
-//!   Not every node has this role.
 
-pub mod attester;
 pub mod node;
 pub mod proto;
 pub mod validator;
