+++ node/libs/roles/src/validator/keys/mod.rs
@@ -1,6 +1,7 @@
 //! Cryptographic keys representing the validator role.
 
 mod aggregate_signature;
+mod proof_of_possession;
 mod public_key;
 mod secret_key;
 mod signature;
@@ -9,6 +10,7 @@ mod testonly;
 mod tests;
 
 pub use aggregate_signature::AggregateSignature;
+pub use proof_of_possession::ProofOfPossession;
 pub use public_key::PublicKey;
 pub use secret_key::SecretKey;
-pub use signature::{ProofOfPossession, Signature};
+pub use signature::Signature;
