+++ node/tools/src/bin/keys.rs
@@ -3,18 +3,17 @@
 
 use crypto::TextFmt as _;
 use zksync_consensus_crypto as crypto;
-use zksync_consensus_roles::{attester, node, validator};
+use zksync_consensus_roles::{node, validator};
 
-/// This tool generates a validator/attester/node key pair and prints it to stdout.
+/// This tool generates a validator/node key pair (including the proof of possession
+/// for the validator key) and prints it to stdout.
 fn main() {
     let validator_key = validator::SecretKey::generate();
-    let attester_key = attester::SecretKey::generate();
     let node_key = node::SecretKey::generate();
     println!("keys:");
     println!("{}", validator_key.encode());
     println!("{}", validator_key.public().encode());
-    println!("{}", attester_key.encode());
-    println!("{}", attester_key.public().encode());
+    println!("{}", validator_key.sign_pop().encode());
     println!("{}", node_key.encode());
     println!("{}", node_key.public().encode());
 }
